Refactored and added unit tests to completely and more precicely test legacy code.
Refactored test "is only valid with only name specified"
Refactored test "minimum" name to a more explicit name.
Refactored test "a complete ingredient object" for a more explicit name.
Refactored test "it allows a valid ingredient to be added. Added expect function
to ask for the object under test, not only the size of ingredients.